:hammer:Update time to long

+++ openjob-common/src/main/java/io/openjob/common/request/WorkerDelayAddRequest.java
@@ -45,5 +45,5 @@ public class WorkerDelayAddRequest implements Serializable {
     /**
      * Delay task execute time.
      */
-    private Integer executeTime;
+    private Long executeTime;
 }
